캐릭터 움직이면 떨리는 현상 없애기

+++ Assets/Scripts/Item/Item.cs
@@ -9,8 +9,8 @@ public class Item: MonoBehaviour
     public int Att;
     public int Def;
     public int MaxHP;
-    public int MovSpd;
-    public int AttSpd;
+    public float MovSpd;
+    public float AttSpd;
     public int Money;
 
     void Awake()
